requirements.txt
setuptools>=68.0.0
wheel>=0.40.0
pip>=23.0.0
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
gunicorn==21.2.0
pytesseract==0.3.10  # Add this line